generalize GalleryDisplay
